BackEnd/admin-billing-service/requirements.txt
Django==5.0.14
django-cors-headers==4.7.0
pytz==2025.2
python-dotenv==1.1.1
PyMySQL==1.1.1
certifi==2025.8.3
reportlab==4.4.4
razorpay==1.4.2
boto3==1.35.0
pymongo==4.7.1